requirements.txt
# Core Agent Zero dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.0
pydantic>=2.5.0
asyncio-mqtt>=0.11.0

# Local SLM training and inference
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
datasets>=2.14.0
bitsandbytes>=0.41.0

# Training optimization (optional)
# unsloth  # For faster training
wandb>=0.16.0  # For experiment tracking

# Remote API clients
openai>=1.3.0
anthropic>=0.7.0
aiohttp>=3.9.0

# Model management
huggingface-hub>=0.19.0

# Utilities
PyYAML>=6.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
mypy>=1.6.0

# Existing dependencies
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
kiwisolver==1.4.8
matplotlib==3.10.3
munkres==1.1.4
packaging==25.0
pillow==11.3.0
pip==25.1.1
pyparsing==3.2.3
PySide6==6.9.1
python-dateutil==2.9.0.post0
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
tornado==6.5.1
unicodedata2==16.0.0
wheel==0.45.1